requirements.txt
asgiref==3.5.2
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dbfread==2.0.7
Django==4.1.3
joblib==1.2.0
lxml==4.9.1
mysqlclient==2.1.1
nltk==3.7
pdf2image==1.16.0
pdfminer.six==20221105
Pillow==9.3.0
pycparser==2.21
pycryptodome==3.16.0
PyPDF2==2.11.2
PySastrawi==1.2.0
pytz==2022.6
regex==2022.10.31
singledispatch==3.7.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
Whoosh==2.7.4
xlrd==2.0.1
